Events, Better disposing/child objects, better bgm position setting.
Boss aura will now automatically spawn on normal bosses, special
spawning bosses need to be replaced.
Objects can now contain child objects which will be deleted with the
parent once it is removed, allowing effects to stick to a boss/enemy as
long as they live, even when the stage calls for a clear.
BGM positioning and disposing has been moved to another thread, making
it not lag the game.